LayoutTests/fast/canvas/webgl/drawArraysOutOfBounds.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css"/>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>
description("Test of drawArrays with out-of-bounds parameters");

var context = create3DDebugContext();
var program = loadStandardProgram(context);

context.useProgram(program);
var vertexObject = context.createBuffer();
context.bindBuffer(context.ARRAY_BUFFER, vertexObject);
context.enableVertexAttribArray(0);

debug("Test empty buffer")
context.bufferData(context.ARRAY_BUFFER, new WebGLFloatArray([  ]), context.STATIC_DRAW);
context.vertexAttribPointer(0, 3, context.FLOAT, false, 0, 0);
shouldThrow("context.drawArrays(context.TRIANGLES, 0, 1)");
shouldThrow("context.drawArrays(context.TRIANGLES, 0, 10000)");
shouldThrow("context.drawArrays(context.TRIANGLES, 0, 10000000000000)");
shouldThrow("context.drawArrays(context.TRIANGLES, 0, -1)");
shouldThrow("context.drawArrays(context.TRIANGLES, 1, 0)");
shouldThrow("context.drawArrays(context.TRIANGLES, -1, 0)");
shouldThrow("context.drawArrays(context.TRIANGLES, 1, -1)");
shouldThrow("context.drawArrays(context.TRIANGLES, -1, 1)");

debug("")
debug("Test buffer with 3 float vectors")
context.bufferData(context.ARRAY_BUFFER, new WebGLFloatArray([ 0,0.5,0, -0.5,-0.5,0, 0.5,-0.5,0 ]), context.STATIC_DRAW);
context.vertexAttribPointer(0, 3, context.FLOAT, false, 0, 0);
shouldBe("context.checkFramebufferStatus(context.FRAMEBUFFER)", "context.FRAMEBUFFER_COMPLETE");
shouldBe("context.drawArrays(context.TRIANGLES, 0, 3)", "undefined");
shouldThrow("context.drawArrays(context.TRIANGLES, 3, 2)");
shouldThrow("context.drawArrays(context.TRIANGLES, 0, 10000)");
shouldThrow("context.drawArrays(context.TRIANGLES, 0, 10000000000000)");
shouldThrow("context.drawArrays(context.TRIANGLES, 0, -1)");
shouldThrow("context.drawArrays(context.TRIANGLES, -1, 0)");
shouldThrow("context.drawArrays(context.TRIANGLES, 1, -1)");
shouldThrow("context.drawArrays(context.TRIANGLES, -1, 1)");

debug("")
successfullyParsed = true;
</script>

<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
