LayoutTests/fast/canvas/webgl/drawArraysOutOfBounds-expected.txt

Test of drawArrays with out-of-bounds parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test empty buffer
PASS context.drawArrays(context.TRIANGLES, 0, 1) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, 0, 10000) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, 0, 10000000000000) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, 0, -1) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, 1, 0) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, -1, 0) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, 1, -1) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, -1, 1) threw exception GL error 1282 in drawArrays.

Test buffer with 3 float vectors
PASS context.checkFramebufferStatus(context.FRAMEBUFFER) is context.FRAMEBUFFER_COMPLETE
PASS context.drawArrays(context.TRIANGLES, 0, 3) is undefined
PASS context.drawArrays(context.TRIANGLES, 3, 2) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, 0, 10000) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, 0, 10000000000000) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, 0, -1) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, -1, 0) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, 1, -1) threw exception GL error 1282 in drawArrays.
PASS context.drawArrays(context.TRIANGLES, -1, 1) threw exception GL error 1282 in drawArrays.

PASS successfullyParsed is true

TEST COMPLETE
